Added threads variable, merged files, added comments
