+++ World.go
@@ -3,6 +3,7 @@ package main
 import (
 	"math/rand"
 	"sync"
+	"time"
 )
 
 type World struct {
@@ -13,6 +14,7 @@ type World struct {
 	fishBreed  int
 	sharkBreed int
 	starve     int
+	threads    int
 }
 
 /**
@@ -28,7 +30,7 @@ type World struct {
 * @param gridSize		An int array of size 2, position 0 holds the width of the grid, position 1 holds the height
 * @return A pointer towards a newly created struct of type World
  */
-func newWorld(numShark, numFish, fishBreed, sharkBreed, starve int, gridSize [2]int) *World {
+func newWorld(numShark, numFish, fishBreed, sharkBreed, starve int, gridSize [2]int, threads int) *World {
 	width := gridSize[0]
 	height := gridSize[1]
 	w := &World{
@@ -37,6 +39,7 @@ func newWorld(numShark, numFish, fishBreed, sharkBreed, starve int, gridSize [2]
 		fishBreed:  fishBreed,
 		sharkBreed: sharkBreed,
 		starve:     starve,
+		threads:    threads,
 	}
 	w.fillTheGrid()
 	w.populateWorld(numFish, numShark)
@@ -94,7 +97,7 @@ func (w *World) spawnCreature(creatureId, x, y int, mutex *sync.Mutex) {
 		creatureId, x, y,
 		w.starve,
 		breedTime)
-	mutex.Lock() // FIX THIS !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+	mutex.Lock()
 	w.creatures = append(w.creatures, creature)
 	w.grid[x][y] = creature
 	mutex.Unlock()
@@ -147,68 +150,148 @@ func (w *World) placeCreatures(ncreatures, creatureId int) {
 * @return 		Returns a pointer array of 4 of the direct neighbours of a given creature
  */
 //////////////////////////////////
-func (w *World) getAndBlockNeighbours(x, y int) (neighbours []*Creature) {
-	for _, movement := range movements {
+func (w *World) getNeighbours(x, y int) (neighbours [4]*Creature) {
+	for i, movement := range movements {
 		directionX := movement[0]
 		directionY := movement[1]
 		newX := (x + directionX + w.width) % w.width
 		newY := (y + directionY + w.height) % w.height
-		select {
-		case w.grid[newX][newY].usedChan <- true:
-			neighbours = append(neighbours, w.grid[newX][newY])
-		default:
-			// Nothing happens
-		}
+
+		neighbours[i] = w.grid[newX][newY]
 	}
 	return neighbours
 }
-func (w *World) moveCreatures(creature *Creature, semChannel chan bool, mutex *sync.Mutex, wg *sync.WaitGroup) {
-	select {
-	case creature.usedChan <- true:
-		neighbours := w.getAndBlockNeighbours(creature.x, creature.y)
-		creature.fertility += 1
-		moved := false
-		x := creature.x
-		y := creature.y
-		var neighbour *Creature
-		if creature.id == 2 {
-			if checkIfAnyNeighbourIsFood(neighbours) {
-				neighbours := getFoodNeighbours(neighbours)
-				neighbour = randomiseNeighbour(neighbours)
-				creature.energy = w.starve
-				w.grid[neighbour.x][neighbour.y].dead = true
-				moved = true
-			}
+
+/**
+* @brief Gets neighbouring positions that contain fish
+*
+* An array containing Creature pointers to the 4 neighbours of a given position are checked to see whether they have any fish present.
+* If a fish is present they are added to a slice. Then that fish slice is returned.
+*
+* @param neighbours 	An Creature pointer array of all the 4 neighbours
+* @return 				Returns a Creature pointers slice of neighbouring fish.
+ */
+func getFoodNeighbours(neighbours [4]*Creature) []*Creature {
+	var neighbour []*Creature
+	for i := 0; i < len(neighbours); i++ {
+		if neighbours[i].id == 1 {
+			neighbour = append(neighbour, neighbours[i])
 		}
-		if !moved {
-			emptyNeighbours := getEmptyNeighbours(neighbours)
-			if len(emptyNeighbours) != 0 {
-				neighbour = randomiseNeighbour(emptyNeighbours)
-				if creature.id == 2 {
-					creature.energy--
+	}
+	return neighbour
+}
+
+/**
+* @brief Chooses a random position
+*
+* A slice containing between 1 and 4 Creature pointers is passed in.
+* From that array 1 random neighbour is selected and returned.
+*
+* @param neighbours 	A slice containing Creature pointers containing available positions
+* @return 				Returns a pointer to a Creature struct
+ */
+
+func randomiseNeighbour(neighbours []*Creature) (neighbour *Creature) {
+	neighbour = nil // If no neighbour present thats not being used it will be set to nil
+	start := rand.Intn(len(neighbours))
+	// Below selects a free random neighbour
+Loop:
+	for i := 0; i < len(neighbours); i++ {
+		select {
+		case neighbours[start].usedChan <- true:
+			neighbour = neighbours[start]
+			break Loop
+		default:
+			start = (start + 1) % len(neighbours)
+		}
+	}
+	return neighbour
+
+}
+
+/**
+* @brief Gets pointers to empty neighbours
+*
+* 4 Direct neighbouring positions(North, West, South and East) are passed into this function.
+* The the ones that don't contain a fish or shark are returned
+*
+* @param neighbours 	An array with 4 Creature pointers in positions that are neighbouring a given creature.
+* @return 				Returns a slice containing pointers to empty creatures from the neighbours array.
+ */
+func getEmptyNeighbours(neighbours [4]*Creature) (emptyNeighbours []*Creature) {
+	for i := 0; i < len(neighbours); i++ {
+		if neighbours[i].id == 0 {
+			emptyNeighbours = append(emptyNeighbours, neighbours[i])
+		}
+	}
+	return emptyNeighbours
+}
+
+/**
+* @brief iterates all the creatures present on the creature slice
+*
+* 4 Direct neighbouring positions(North, West, South and East) are passed into this function.
+* The the ones that don't contain a fish or shark are returned
+*
+* @param neighbours 	An array with 4 Creature pointers in positions that are neighbouring a given creature.
+ */
+func (w *World) iterateCreatures(creature *Creature, semChannel chan bool, mutex *sync.Mutex, wg *sync.WaitGroup) {
+	neighbours := w.getNeighbours(creature.x, creature.y)             // Gets 4 nearby grid neighbours
+	if !creature.dead && creature == w.grid[creature.x][creature.y] { // If creature dead or no longer on grid just skip this function
+		select {
+		case creature.usedChan <- true: // If creature is free
+			creature.fertility += 1 // Plus 1 chronon towards breeding timer
+			moved := false
+			x := creature.x
+			y := creature.y
+			var neighbour *Creature
+			if creature.id == 2 {
+				foodNeighbours := getFoodNeighbours(neighbours) // Checks if shark can eat
+				if len(foodNeighbours) != 0 {
+					neighbour = randomiseNeighbour(foodNeighbours)
+					if neighbour != nil {
+						// Eat Fish below
+						creature.energy = w.starve
+						w.grid[neighbour.x][neighbour.y].dead = true
+						moved = true
+					}
 				}
-				moved = true
 			}
-		}
-		if creature.energy < 0 {
-			creature.dead = true
-			w.grid[x][y] = newCreatureEmpty(x, y)
-			<-neighbour.usedChan
-		} else if moved {
-			creature.x = neighbour.x
-			creature.y = neighbour.y
-			w.grid[neighbour.x][neighbour.y] = creature
-			if creature.fertility >= creature.breedTime {
-				creature.fertility = 0
-				w.spawnCreature(creature.id, x, y, mutex)
-			} else {
+			if !moved {
+				emptyNeighbours := getEmptyNeighbours(neighbours)
+				if len(emptyNeighbours) != 0 {
+					neighbour = randomiseNeighbour(emptyNeighbours)
+					if creature.id == 2 {
+						creature.energy--
+					}
+					if neighbour != nil {
+						moved = true
+					}
+				}
+			}
+			// Checks if sharks going to starve
+			if creature.energy < 0 {
+				creature.dead = true
 				w.grid[x][y] = newCreatureEmpty(x, y)
+				if neighbour != nil {
+					<-neighbour.usedChan // Shark dies so free neighbour
+				}
+			} else if moved {
+				creature.x = neighbour.x
+				creature.y = neighbour.y
+				w.grid[neighbour.x][neighbour.y] = creature
+				// Check if creature can breed
+				if creature.fertility >= creature.breedTime {
+					creature.fertility = 0
+					w.spawnCreature(creature.id, x, y, mutex)
+				} else {
+					w.grid[x][y] = newCreatureEmpty(x, y)
+				}
+				<-neighbour.usedChan
 			}
-			<-neighbour.usedChan
+			<-creature.usedChan
+		case <-time.After(timeout):
 		}
-		<-creature.usedChan
-	default:
-		// Do nothing
 	}
 	wg.Done()
 	<-semChannel
@@ -227,10 +310,10 @@ func (w *World) iterateProgram() {
 	rand.Shuffle(len(w.creatures),
 		func(i, j int) { w.creatures[i], w.creatures[j] = w.creatures[j], w.creatures[i] })
 
-	// Works with 1 so far
-	semChannel := make(chan bool, 8)
+	semChannel := make(chan bool, w.threads)
 	var mutex sync.Mutex
 	var wg sync.WaitGroup
+	// Since creatures slice grows during execution I needed a way to keep track the max for this itteration
 	ncreatures := len(w.creatures)
 	for i := 0; i < ncreatures; i++ {
 		creature := w.creatures[i]
@@ -249,86 +332,4 @@ func (w *World) iterateProgram() {
 		}
 	}
 	w.creatures = newCreatures
-}
-
-/**
-* @brief Gets pointers to empty neighbours
-*
-* 4 Direct neighbouring positions(North, West, South and East) are passed into this function.
-* The the ones that don't contain a fish or shark are returned
-*
-* @param neighbours 	An array with 4 Creature pointers in positions that are neighbouring a given creature.
-* @return 				Returns a slice containing pointers to empty creatures from the neighbours array.
- */
-func getEmptyNeighbours(neighbours []*Creature) (emptyNeighbours []*Creature) {
-	for i := 0; i < len(neighbours); i++ {
-		if neighbours[i].id == 0 {
-			emptyNeighbours = append(emptyNeighbours, neighbours[i])
-		} else {
-			<-neighbours[i].usedChan
-			//print("Released 4, ")
-		}
-	}
-	return emptyNeighbours
-}
-
-/**
-* @brief Chooses a random position
-*
-* A slice containing between 1 and 4 Creature pointers is passed in.
-* From that array 1 random neighbour is selected and returned.
-*
-* @param neighbours 	A slice containing Creature pointers containing available positions
-* @return 				Returns a pointer to a Creature struct
- */
-func randomiseNeighbour(neighbours []*Creature) *Creature {
-	neighbour := neighbours[rand.Intn(len(neighbours))]
-	for i := 0; i < len(neighbours); i++ {
-		if neighbour != neighbours[i] {
-			<-neighbours[i].usedChan
-			//print("Released 5, ")
-		}
-	}
-	return neighbour
-}
-
-/**
-* @brief Checks whether any of the neighbours have a fish inside.
-*
-* An array containing Creature pointers to the 4 neighbours of a given position are checked to see whether they have any fish present.
-* If a fish is present true is returned. This function is only used by sharks.
-*
-* @param neighbours 	An Creature pointer array of all the 4 neighbours
-* @return 				Returns a boolean
- */
-func checkIfAnyNeighbourIsFood(neighbours []*Creature) bool {
-	foodPresent := false
-	for i := 0; i < len(neighbours); i++ {
-		if neighbours[i].id == 1 {
-			foodPresent = true
-		}
-	}
-	return foodPresent
-}
-
-/**
-* @brief Gets neighbouring positions that contain fish
-*
-* An array containing Creature pointers to the 4 neighbours of a given position are checked to see whether they have any fish present.
-* If a fish is present they are added to a slice. Then that fish slice is returned.
-*
-* @param neighbours 	An Creature pointer array of all the 4 neighbours
-* @return 				Returns a Creature pointers slice of neighbouring fish.
- */
-func getFoodNeighbours(neighbours []*Creature) []*Creature {
-	var neighbour []*Creature
-	for i := 0; i < len(neighbours); i++ {
-		if neighbours[i].id == 1 {
-			neighbour = append(neighbour, neighbours[i])
-		} else {
-			<-neighbours[i].usedChan
-			//print("Released 6, ")
-		}
-	}
-	return neighbour
 }
